Feat: add folder layout + added components folder + added h1 in index page
