+++ components/Layout/layout.tsx
@@ -1,0 +1,16 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+
+
+const inter = Inter({ subsets: ["latin"] });
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <div className={inter.className}>{children}</div>
+  );
+}
+
