+++ pages/_app.js
@@ -1,0 +1,17 @@
+import { Inter } from "next/font/google";
+import "../public/styles/globals.css";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export default function MyApp({ Component, pageProps }) {
+  return (
+    <>
+      <style jsx global>{`
+        html {
+          font-family: ${inter.style.fontFamily};
+        }
+      `}</style>
+      <Component {...pageProps} />
+    </>
+  );
+}
