--- src/app/page.tsx
@@ -1,12 +1,0 @@
-import Image from "next/image";
-import styles from "./page.module.css";
-
-export default function Home() {
-  return (
-    <main>
-      <section>
-        <h1>Helloworld</h1>
-      </section>
-      </main>
-  );
-}
